# internal/http/openapi_spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GoClaw Gateway API",
    "description": "PostgreSQL multi-tenant AI agent gateway with WebSocket RPC + HTTP API.\n\n## Authentication\n\nAll endpoints require a Bearer token in the `Authorization` header:\n\n```\nAuthorization: Bearer <gateway-token-or-api-key>\n```\n\nYou can use either the **gateway token** (grants admin access) or an **API key** created via the API Keys endpoints (grants scoped access).\n\nIf no token is configured on the server, authentication is disabled (backward compatibility).\n\n## Common Headers\n\n| Header | Description |\n|--------|-------------|\n| `X-GoClaw-User-Id` | External user ID for multi-tenant context |\n| `X-GoClaw-Agent-Id` | Target agent ID (alternative to model prefix) |\n| `X-GoClaw-Tenant-Id` | Tenant scope — UUID or slug (gateway token / cross-tenant API keys) |\n| `Accept-Language` | Locale for error messages (`en`, `vi`, `zh`) |\n\n## WebSocket Protocol\n\nConnect via `POST /ws` (upgrade). Protocol v3 uses frame types: `req`, `res`, `event`.\nFirst request must be `connect` with `{\"token\": \"...\", \"user_id\": \"...\", \"locale\": \"en\"}`.",
    "version": "0.2.0",
    "contact": {
      "name": "GoClaw",
      "url": "https://github.com/nextlevelbuilder/goclaw"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/v1/storage/files": {
      "get": {
        "tags": ["Storage"],
        "summary": "List workspace files",
        "parameters": [
          { "name": "path", "in": "query", "schema": { "type": "string" }, "description": "Subdirectory path" }
        ],
        "responses": { "200": { "description": "File listing" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Gateway token or API key"
      }
    },
...
  }
}